fix: 将 NewSnowflake 的 logicalId 参数类型从 uint 改为 int，并修正测试中的类型定义

+++ util/snowflake_test.go
@@ -22,7 +22,7 @@ func newTestSnowflake(baseTime time.Time, logicalId int64, logicalBits uint, cou
 func TestNewSnowflakeLogicalIdBoundaries(t *testing.T) {
 	// 用于与 NewSnowflake 一致的常量
 	const logicalIdBits uint = 13
-	max := uint(1<<logicalIdBits - 1)
+	max := 1<<logicalIdBits - 1
 
 	// 合法
 	if _, err := NewSnowflake(time.Now().Add(-time.Second), 0); err != nil {
